Add a little C setuid programm to clear the nginx cache
Nginx stores the cache files with 700 mode, which doesn't
allow the php process to remove the cache if run under another
user than nginx.

The programm need to be compiled, use the provided makefile.
Note: You'll be prompted for a sudo password to be able to set
the setuid bit and the owner.