purge: Migrate from a setuid program to a nginx perl handler
The main motivation for this change is, that we do not
need a compiler suite on the server anymore.
We could keep the setuid program for people that prefer that,
but that's probably not worth the maintainance costs..
(and it can be checked out from git, if ever needed)

A note on why we use perl: The lua nginx module is not part of
NGINX core. Perl is part of the core system and (at least on Fedora)
installable without manual compiling.

Furthermore the nnginx server can now be placed on a
different host than the TYPO3 system.